orf radiothek: renamed classes to helper, removed catchup day conf entry set it to 30 days flat
